Init test with vitest
