+++ proxy/package.json
@@ -7,7 +7,7 @@
     "dev": "tsx watch index.ts",
     "build": "tsc",
     "start": "node .dist/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "vitest --ui --coverage.enabled=true"
   },
   "author": "",
   "dependencies": {
@@ -28,8 +28,11 @@
   "devDependencies": {
     "@types/express": "^5.0.0",
     "@types/node": "^22.15.31",
+    "@vitest/coverage-v8": "3.2.4",
+    "@vitest/ui": "3.2.4",
     "nodemon": "^3.1.7",
     "tsx": "^4.19.2",
-    "typescript": "^5.6.3"
+    "typescript": "^5.6.3",
+    "vitest": "^3.2.4"
   }
 }
